+++ docs/CHANGELOG.rst
@@ -18,6 +18,9 @@ Changed
 -------
 - Use ``resolwebio/wgbs:3.0.0`` in ``walt``, ``methcounts``, ``hmr`` and
   ``bs-conversion-rate`` processes
+- Use ``resolwebio/chipseq:6.0.0`` in ``chipqc``, ``macs2-callpeak``,
+  ``macs14``, ``qc-prepeak``, ``chipseq-peakscore``,
+  ``chipseq-genescore`` and ``upload-bed`` processes
 
 Fixed
 -----